StateV2: Deprecate various combinators
This commit deprecates various combinators that are no longer necessary.

It also deprecates `map` and `bimap` because both use `memo` by default
when this is often not necessary. Replacements which make this behavior
explicit are provided for both.
The `map` replacement additionally now also provides access to the
`Observer`, which allows observing additional states in that lambda.

Where possible with zero chance of breaking existing code, these
replacements are then used in other combinators.

Source-Commit: 99334e0d5c1200bbdb76e5d8a919cf45e3c234aa