StateV2: Fix `effect` not receiving updates if triggered during its own execution
This was because we'd only mark the node as `Clean` after the update
function had completed. And therefore any changes made during the
execution of the update function wouldn't re-queue the node because it
was already marked as `Dirty`.

This commit fixes this issue by marking the node as `Clean` right before
we invoke the update function. This then allows it to be re-queued in
case a change happens.

An additional fix to the `getTracked` method is required to prevent that
function from also re-marking the node as Dirty on initial subscription
when the dependency node happened to be dirty at the time. See its code
comment for details.

Source-Commit: e10413c1b35e041d1fad5e52754df81b2d8395f7